Create the main plugin file in `.plugin`
This is more future-proof for plugins that grow additional submodules to
handle different things. For example, a plugin with lots of helper
functions that its author decides would be better in their own `.util`
submodule will run into circular-import problems if the main plugin
lives in `__init__.py` and tries to `from . import util`, but the same
import works just fine from `plugin.py`.
